Revert "Chore: Consolidate hooks"
This reverts commit af2356b0a263acd3be5218c4414dbe24b8015b75.
