+++ components/Headers/AppHeader.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
 import { COLORS } from '../../helpers/constants'
-import { useFCL } from '../../hooks'
+import { useFCL } from '../../hooks/useFCL'
 
 const HeaderContainer = styled.div`
   display: flex;
