--- hooks/index.js
@@ -1,4 +1,0 @@
-import useFCL from './useFCL'
-import useLocalStorage from './useLocalStorage'
-
-export { useFCL, useLocalStorage }
